requirements.txt
fastapi
google-genai
google-generativeai
python-dotenv
chromadb
psycopg2
requests
numpy
streamlit